+++ actionControllers/BusinessUnitController.ts
@@ -46,11 +46,9 @@ export const getMe: ActionHook = async (request: Request, actionContext: ActionC
 
 export const setMe: ActionHook = async (request: Request, actionContext: ActionContext) => {
   const businessUnitApi = new BusinessUnitApi(actionContext.frontasticContext, getLocale(request));
-  const storeApi = new StoreApi(actionContext.frontasticContext, getLocale(request));
   const data = JSON.parse(request.body);
 
   const businessUnit = await businessUnitApi.get(data.key, request.sessionData?.account?.accountId);
-  const store = businessUnit.stores?.[0]?.key ? await storeApi.get(businessUnit.stores[0].key) : undefined;
   const organization = await businessUnitApi.getOrganizationByBusinessUnit(businessUnit);
   const response: Response = {
     statusCode: 200,