+++ mappers/StoreMappers.ts
@@ -7,8 +7,21 @@ export class StoreMappers {
     locale: string,
   ): Store {
     return {
-      ...store,
       name: store.name?.[locale],
+      id: store.id,
+      key: store.key,
+      distributionChannels: store.distributionChannels,
+      supplyChannels: store.supplyChannels,
+    };
+  }
+  static mapStoreToSmallerStore(
+    store: CommercetoolsStore,
+    locale: string,
+  ): Store {
+    return {
+      name: store.name?.[locale],
+      id: store.id,
+      key: store.key,
     };
   }
 }